feat: Change the syntax of the front end
And make it more flexible: we take in a term of the form
(fun x y z .. => exists w, P x y w) and run the elaborator to get
somewhat more robust parsing. We now error if an argument to P is
not a variable, which we probably wanted to do anyway.